UI UPDATED

+++ src/components/products/ProductItem.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
+import Swal from 'sweetalert2';
 
 import { addToCart, deleteOneProduct, productSetActive } from "../../actions/cart";
 import { uiOpenModal } from '../../actions/ui';
@@ -20,7 +21,24 @@ export const ProductItem = ({data}) => {
     }
 
     const deleteProduct= (id) => {
-        dispatch( deleteOneProduct(id) );
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    dispatch( deleteOneProduct(id) );
+                    Swal.fire(
+                    'Deleted!',
+                    'Your product has been deleted.',
+                    'success'
+                    )
+                }
+            })
     }
 
     return (
